requirements.txt
fastapi
mysql-connector-python
uvicorn[stadard]
sqlalchemy
passlib
pyjwt
python-multipart
